Add robust bad ACE data handling and test scenarios
Instead of dropping bad ACE data at the point of reading in
values, leave the bad values there (as indicated by a negative
value) and do downstream handling of bad values.

Indicate bad or N/A values with bold red text in the state table.
If there is no average P3 from MTA then use 0.0 for fluence but
show fluence as red.